% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{burritos_cost}
\alias{burritos_cost}
\title{Calculate total burrito price}
\usage{
burritos_cost(price, count)
}
\arguments{
\item{price}{price per burrito ($)}

\item{count}{number of burritos in order}
}
\description{
Calculate total burrito price
}
