% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicken_eggs.R
\name{chicken_eggs}
\alias{chicken_eggs}
\title{Number of eggs you get from your chickens}
\usage{
chicken_eggs(chickens, eggs)
}
\arguments{
\item{chickens}{number of chickens}

\item{eggs}{number of eggs each chicken lays}
}
\description{
Number of eggs you get from your chickens
}
